+++ R/downloadParseTimeseriesData.R
@@ -46,6 +46,13 @@ downloadParseTimeseriesData <- function(
       dplyr::pull(.data$timezone)
   }
   
+  if ( length(timezone) > 1 ) {
+    err_msg <- paste0(length(timezone),
+                      " senors share the label '",
+                      label, "'")
+    stop(err_msg)
+  }
+  
   # Create a valid dateRange
   if ( !is.null(startdate) && !is.null(enddate) ) {
     # Don't require day boundaries