+++ R/pat_outliers.R
@@ -164,8 +164,8 @@ pat_outliers <- function(
     
     # Set time axis to sensor local time
     timezone <- pat$meta$timezone
-    A_flagged$datetime <- lubridate::with_tz(A_flagged$datetime, timezone)
-    B_flagged$datetime <- lubridate::with_tz(B_flagged$datetime, timezone)
+    A_flagged$datetime <- lubridate::with_tz(A_flagged$datetime, tzone = timezone)
+    B_flagged$datetime <- lubridate::with_tz(B_flagged$datetime, tzone = timezone)
     
     # Use the same y limits for both plots
     ylim <- range(c(A_data$pm25_A, B_data$pm25_B), na.rm = TRUE)